End_to_End_code/requirements.txt
gradio>=4.44.0
torch>=2.2.0
torchvision>=0.17.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
